refactor: step4 로또(수동)
controller에서 view로직 분리
